Allow parsing from rewritten nonterminals (#46)
This PR allows parsing from the nonterminals that will get rewritten during the datadependent grammar transformations. This is particularly useful to start parsing from expression grammar nonterminals. In this PR we generate nonterminals of the form `$_expression = expression(0)` and provide mechanism to pick up these top-level nonterminals when parsing from `expression` and also convert the parse trees back to the expected one (no `$_expression` node is present in the final parse tree).