Panic mode error recovery (#48)
This PR adds the first version of an error recovery mechanism for Iguana. The error recovery mechanism is similar to the panic mode error recovery found in recursive descent parsers. We add the special `error` symbol to the grammar that provides synchronization points. When a parse error occurs, we empty the stack (by traversing the GSS) to find the synchronization points and continue parsing from there. Given the non-deterministic nature of GLL, the error recovery should happen after parsing is finished.